findGlobals {codetools}	R Documentation

Find Global Functions and Variables Used by a Closure

Description

Finds global functions and variables used by a closure.

Usage

findGlobals(fun, merge = TRUE) 

Arguments

fun 	 closure.

merge 	 logical

Details

The result is an approximation. R semantics only allow variables that might be local to be identified (and event that assumes no use of assign and rm).

Value

Character vector if merge is true; otherwise, a list with functions and variables components.

Author(s)

Luke Tierney

Examples

findGlobals(findGlobals)
findGlobals(findGlobals, merge = FALSE)

[Package codetools version 0.2-2 Index]